# api-gateway/swagger.yaml
openapi: 3.0.0
info:
  title: Photo Prestiges API
  description: Microservices architectuur voor Photo Prestiges
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: api-gateway
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      summary: Gebruiker registreren
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum: [participant, owner]
      responses:
        '201':
          description: Gebruiker geregistreerd
        '500':
          description: Gebruiker registreren mislukt
  /auth/login:
    post:
      summary: Gebruiker inloggen
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Gebruiker ingelogd + JWT token
        '401':
          description: Ongeldige gegevens
        '500':
          description: Gebruiker inloggen mislukt
  /targets:
    post:
      summary: Target aanmaken (owners)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                locationDescription:
                  type: string
                deadline:
                  type: string
                  format: date-time
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Target aangemaakt
        '400':
          description: Afbeelding is verplicht
        '403':
          description: Niet geauthoriseerd om target aan te maken
  /targets/{id}:
    delete:
      summary: Target verwijderen (owners)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Target verwijderd
        '403':
          description: Niet geauthoriseerd om target te verwijderen
        '404':
          description: Target niet gevonden
        '500':
          description: Target verwijderen mislukt
  /explore/active-targets:
    get:
      summary: Openstaande targets ophalen
      parameters:
        - in: query
          name: location
          schema:
            type: string
          description: Filter by location description
      responses:
        '200':
          description: Lijst van actieve targets
        '500':
          description: api-gateway routing error
  /participate/{targetId}:
    post:
      summary: Entry indienen (participants)
      parameters:
        - in: path
          name: targetId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Entry ingediend
        '400':
          description: Afbeelding is verplicht
        '403':
          description: Target deadline is voorbij
        '500':
          description: Entry indienen mislukt
  /participate/{id}:
    delete:
      summary: Eigen entry verwijderen
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Entry verwijderd
        '403':
          description: Niet geauthoriseerd om entry te verwijderen
        '404':
          description: Entry niet gevonden
        '500':
          description: Entry verwijderen mislukt
  /participate/{targetId}/my-score:
    get:
      summary: Eigen score van target ophalen
      parameters:
        - in: path
          name: targetId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Score opgehaald
        '404':
          description: Geen entry ingediend
        '500':
          description: Score ophalen mislukt
  /participate/{targetId}/scores:
    get:
      summary: Scores van target ophalen (owners)
      parameters:
        - in: path
          name: targetId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Scores opgehaald
        '403':
          description: Niet geauthoriseerd om scores op te halen